+++ services/backend/src/router/admin/admin.js
@@ -1,6 +1,6 @@
 const { Router } = require("express");
 const { keycloak } = require("../../auth/keycloak");
-const admin = require("../../core/admin/index");
+const admin = require("../../core/admin/admin");
 
 const statistics = require("../../core/statistics");
